Split up commands, move tests next to the command impl
