--- lapce-data/src/editor/test/commands/mod.rs
@@ -1,1 +1,0 @@
-mod insert_tab;
